+++ Sources/BraintreeShopperInsights/BTShopperInsightsAnalytics.swift
@@ -3,12 +3,10 @@ import Foundation
 enum BTShopperInsightsAnalytics {
     
     // MARK: - Merchant Triggered Events
-
-    static let payPalSelected = "shopper-insights:paypal-selected"
-    static let venmoSelected = "shopper-insights:venmo-selected"
-
+      
     static let buttonPresented = "shopper-insights:button-presented"
-    
+    static let buttonSelected = "shopper-insights:button-selected"
+
     // MARK: - SDK Triggered Events
     
     static let recommendedPaymentsStarted = "shopper-insights:get-recommended-payments:started"