+++ Sources/BraintreeShopperInsights/BTShopperInsightsClient.swift
@@ -106,17 +106,16 @@ public class BTShopperInsightsClient {
             shopperSessionID: shopperSessionID
         )
     }
-    
-    /// Call this method when the PayPal button has been selected/tapped by the buyer.
-    /// This method sends analytics to help improve the Shopper Insights feature experience
-    public func sendPayPalSelectedEvent() {
-        apiClient.sendAnalyticsEvent(BTShopperInsightsAnalytics.payPalSelected, shopperSessionID: shopperSessionID)
-    }
 
-    /// Call this method when the Venmo button has been selected/tapped by the buyer.
-    /// This method sends analytics to help improve the Shopper Insights feature experience
-    public func sendVenmoSelectedEvent() {
-        apiClient.sendAnalyticsEvent(BTShopperInsightsAnalytics.venmoSelected, shopperSessionID: shopperSessionID)
+    /// Call this method when a button has been selected/tapped by the buyer.
+    /// This method sends analytics to help improve the Shopper Insights feature experience.
+    /// - Parameter buttonType: Type of button presented - PayPal, Venmo, or Other
+    public func sendSelectedEvent(for buttonType: BTButtonType) {
+        apiClient.sendAnalyticsEvent(
+            BTShopperInsightsAnalytics.buttonSelected,
+            buttonType: buttonType.rawValue,
+            shopperSessionID: shopperSessionID
+        )
     }
 
     /// Indicates whether the PayPal App is installed.